#203 move ViewModelTest into notifications package, refactor loop from iterator-style to foreach-style.
